Refactor Layer methods to return i32 for width and height; simplify tile retrieval in setup function

+++ src/plugins/tiled.rs
@@ -86,19 +86,20 @@ pub struct Layer<'map> {
 }
 
 impl Layer<'_> {
-    pub fn width(&self) -> u32 {
-        self.tiled_tile_layer.width().unwrap()
+    pub fn width(&self) -> i32 {
+        self.tiled_tile_layer.width().unwrap() as i32
     }
 
-    pub fn height(&self) -> u32 {
-        self.tiled_tile_layer.height().unwrap()
+    pub fn height(&self) -> i32 {
+        self.tiled_tile_layer.height().unwrap() as i32
     }
 
     pub fn tile(&self, x: i32, y: i32) -> Option<Tile> {
-        if let Some(tile) = self.tiled_tile_layer.get_tile(x as i32, y as i32) {
+        if let Some(tile) = self.tiled_tile_layer.get_tile(x, y) {
             if let Some(tile_data) = tile.get_tile() {
                 Some(Tile {
                     id: tile.id(),
+                    // Retrieving a custom property named "collision" only if it exists and is a boolean
                     collision: tile_data.properties.get("collision").and_then(|v| {
                         if let tiled_rs::PropertyValue::BoolValue(b) = v {
                             Some(*b)
